refactor the impl of elementwise grad kernel
